+++ Payloads/ServerMessagePayload.cs
@@ -1,0 +1,8 @@
+
+namespace WebAPI.Payloads
+{
+    public class ServerMessagePayload
+    {
+        public string message { get; set; }
+    }
+}