+++ Routes/Devices/ById/PostDevice.cs
@@ -34,7 +34,7 @@ namespace WebAPI.Routes.Devices.ById
             }
             catch
             {
-                e.Context.SendResponse(500, new ErrorPayload()
+                e.Context.SendResponse(400, new ErrorPayload()
                 {
                     message = "Expected body to be DevicePayload."
                 });