+++ Routes/Players/BySteamId/BanPlayer.cs
@@ -34,7 +34,7 @@ namespace WebAPI.Routes.Devices
             }
             catch
             {
-                e.Context.SendResponse(500, new ErrorPayload()
+                e.Context.SendResponse(400, new ErrorPayload()
                 {
                     message = "Expected body to be BanPlayerPayload."
                 });